requirements.txt
amqp==2.6.1
apispec==3.3.2
APScheduler==3.6.3
asgiref==3.3.0
asn1crypto==0.24.0
astroid==2.4.2
attrs==20.3.0
Babel==2.9.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
braintree==3.59.0
cachetools==4.2.1
celery==4.4.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cryptography==2.6.1
defusedxml==0.7.1
Django==3.1.3
django-crispy-forms==1.11.1
django-taggit==1.2.0
dnspython==2.1.0
email-validator==1.1.2
entrypoints==0.3
Flask==1.1.2
Flask-AppBuilder==3.2.2
Flask-Babel==1.0.0
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
flatten-json==0.1.8
flower==0.9.3
google-api-core==1.26.3
google-auth==1.29.0
google-cloud-core==1.6.0
google-cloud-storage==1.37.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
greenlet==1.0.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.10.0
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.3
json-flatten==0.1
jsonschema==3.2.0
keyring==17.1.1
keyrings.alt==3.1.1
kombu==4.6.11
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.1
mccabe==0.6.1
numpy==1.20.2
oauthlib==3.1.0
packaging==20.7
pandas==1.2.4
Pillow==7.0.0
pip-review==1.1.0
postgres==3.0.0
prison==0.1.3
prompt-toolkit==3.0.18
protobuf==3.15.8
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
PyGObject==3.30.4
PyJWT==2.1.0
pylint==2.6.0
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-telegram-bot==13.5
python3-openid==3.2.0
pytz==2020.4
pyxdg==0.25
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
SecretStorage==2.3.1
selenium==3.141.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==4.1.0
soupsieve==2.2.1
SQLAlchemy==1.3.24
SQLAlchemy-Utils==0.36.8
sqlparse==0.4.1
toml==0.10.2
tornado==5.1.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
vine==1.3.0
virtualenv==15.1.0
wcwidth==0.2.5
webdriver-manager==3.3.0
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
zipp==3.4.1
